Show list of industry identifiers in Book Details
Google books has industry identifiers, which includes things like ISBN
and others. I thought it was just ISBN. This means that there will be an
array of items. So I created a component that maps these items to a list
and then call that component in the main class. Has tests.