Create a Book class
Create a Book class that can be passed to create an Book object, check types and that
everything is well formed. This is in its own file. I wanted to use
interfaces and avoid classes, but it seems this is the way to go, since
you can't create an instance of an interface.